extract moveElement function to metabase/lib
